add global API error handling

+++ churchinfo/service/FinancialService.php
@@ -413,9 +413,25 @@ class FinancialService {
     
 	function deletePayment($groupKey) 
     {
-		$sSQL = "DELETE FROM `pledge_plg` WHERE `plg_GroupKey` = '" . $groupKey . "';";
-		RunQuery($sSQL);
-	}
+        $sSQL = "SELECT deposit_dep.dep_Closed FROM pledge_plg 
+        INNER JOIN deposit_dep ON
+        pledge_plg.plg_depID = deposit_dep.dep_ID
+        WHERE
+        pledge_plg.plg_GroupKey = '".$groupKey."'";
+        $rIDepClosed = RunQuery($sSQL);
+        $isDepositClosed = mysql_fetch_row($rIDepClosed)[0];
+        if( ! $isDepositClosed)
+        {            
+            $sSQL = "DELETE FROM `pledge_plg` WHERE `plg_GroupKey` = '" . $groupKey . "';";
+            RunQuery($sSQL);
+
+        }
+        else
+        {
+            throw new Exception("Cannot delete a payment from a closed deposit");
+        }
+        
+    }
 	
 	function getMemberByScanString($sstrnig)
 	{
